backend/requirements.txt
asgiref==3.7.2
Django
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.4
sqlparse==0.4.4
